+++ cypress-shared/integration/shopper-location/2.4-enable-Location-via-browser-popup-order-product.spec.js
@@ -1,23 +1,13 @@
 import { loginViaAPI, preserveCookie } from '../../support/common/support'
 import { orderProductTestCase } from '../../support/shopper-location/common'
-import {
-  UsDetails,
-  location,
-} from '../../support/shopper-location/outputvalidation'
-
-const { country, postalCode } = UsDetails
+import { UsDetails } from '../../support/shopper-location/outputvalidation'
 
 const prefix = 'Enable location'
 
 describe(`${prefix}- via browser popup - Order the Product`, () => {
   loginViaAPI()
 
-  orderProductTestCase(prefix, {
-    lat: location.lat,
-    long: location.long,
-    country,
-    postalCode,
-  })
+  orderProductTestCase(prefix, UsDetails)
 
   preserveCookie()
 })