+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+- Added wipe testcase & improved 2.3 ,2.4 shopper location testcases.
+
 ### Changed
 - Moved gmail to common folder
 - In, commonGraphlValidation() added expect(response.body).to.not.equal('OK')